requirements.txt
playwright>=1.40
langchain>=0.2
langchain-openai>=0.1
langgraph>=0.1
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
pandas>=2.0
openpyxl>=3.0
pytest>=7.0
